+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Allow changing the event class using Event#with
 
 ## [0.16.1] - 2018-01-17
 - Fixed bug with Sequel gem expecting processes_event_types to be an Array